+++ site/functions/Scaffolding/Changelog.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Changelog from 'public/Changelog'
-import html from 'generator/CHANGELOG.md'
+import html from 'scaffolding/CHANGELOG.md'
 
 export default () => {
   return <Changelog html={html} />