+++ site/functions/Scaffolding/Docs.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Markdown from 'public/Markdown'
-import html from 'generator/DOCS.md'
+import html from 'scaffolding/DOCS.md'
 
 export default () => {
   return <Markdown html={html} />