+++ site/functions/Scaffolding/Workflow.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Markdown from 'public/Markdown'
-import html from 'generator/WORKFLOW.md'
+import html from 'scaffolding/WORKFLOW.md'
 
 export default () => {
   return <Markdown html={html} />